requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
openai-whisper==20231117
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
Pillow==10.1.0
moviepy==1.0.3